backend/requirements.txt
Flask==1.1.2
SQLAlchemy==1.3.18
flask-restful
Flask-SQLAlchemy
psycopg2-binary
flask-cors
SQLAlchemy-Searchable
python-dotenv
mypy
coverage
black